Fixed many of the warnings. Started a little cleanup in the new ACPI/APIC/SMP/Multiprocessor stuff. Increased the x64 stack size. Still trying to fix a bug that overwrites out heap.

+++ src/system/kernel/debugger/commands/DisassembleCommand.cc
@@ -74,7 +74,7 @@ bool DisassembleCommand::execute(const HugeStaticString &input, HugeStaticString
 #endif
   disassembler.setLocation(address);
 
-  for (int i = 0; i < nInstructions; i++)
+  for (size_t i = 0; i < nInstructions; i++)
   {
     text.clear();
     uintptr_t location = disassembler.getLocation();
